Show all lines, not just ones with profiling information
I didn't notice before, but the line profile output will skip lines that it gets
no profiling data for (so you would see a jump in line number in the middle of
the profiling information).

Instead of that, display all the code, with the cells corresonding to cells with
no profiling information blanked out.